nnunet/pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "nnunet-project"
version = "0.1.0"
description = "nnU-Net segmentation project"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "numpy>=1.20.0",
    "pillow>=9.0.0",
    "geojson>=2.5.0",
    "argparse>=1.4.0",
]

[tool.setuptools]
packages = ["nnUNet"]

[project.scripts]
prepare_dataset_for_nnunet = "nnunet.prepare_dataset_for_nnunet:main"

[tool.uv.sources]
nnunetv2 = { path = "nnUNet", editable = true }
